View interactions fixes

+++ src/main/java/server/controller/MemoryManager.java
@@ -14,57 +14,52 @@ public class MemoryManager {
     private ServerModel model;
     private String  dirpath= "./src/main/java/server/accounts";
     private ArrayList<ObjString> accounts;
-    public MemoryManager(ServerModel model) {
+    public MemoryManager(ServerModel model) throws IOException {
 
         this.model=model;
 
         File dir = new File(dirpath);
         if (!dir.isDirectory())
             dir.mkdir();
-        ID_Seq=new IdSequence(dirpath);
+        ID_Seq = new IdSequence(dirpath);
+
         accounts = getAccounts(dir);
-        System.out.println(accounts);
     }
     public ObjString findAccount(String dest){
         return accounts.get(accounts.indexOf(new ObjString(dest)));
     }
 
-    public List<String> addMail(Email mail){
+    public List<String> addMail(Email mail) throws IOException {
         List<String> dests=mail.getReceivers();
         List<String> err=new ArrayList<>();
         for (String dest:dests) {
            if (!accounts.contains(new ObjString(dest) )) err.add(dest);
         }
         if(err.isEmpty()) {
-            System.out.println("sto Inviando");
             mail.setID(ID_Seq.getNextID()); //autoboxing
             save(mail);
             return null;}
-        else return err;
+        else {
+            return err;
+        }
+
     }
 
-    private void save(Email mail) {
+    private void save(Email mail) throws IOException {
         Gson gson=new Gson();
         String path;
         File receiver;
         FileWriter out;
-        System.out.println("sono qua");
         for (String dest:mail.getReceivers()) {
             synchronized (this.findAccount(dest)) {
                 path= dest +"/InArrivo/"+mail.getID();
                 receiver = new File(dirpath+"/"+path);
-                try {
-                    receiver.createNewFile();
+                receiver.createNewFile();
+                out=new FileWriter(receiver) ;
+                out.write(gson.toJson(mail)+"\n");
+                out.flush();
+                out.close();
 
-                    out=new FileWriter(receiver) ;
-                    System.out.println("scrivo "+ gson.toJson(mail));
-                    out.write(gson.toJson(mail)+"\n");
-                    out.flush();
-                    out.close();
-                    model.setLog(model.getLog() + "Email inviata da "+mail.getSender()+" a "+dest+"\n" );
-                } catch (IOException e) {
-                    throw new RuntimeException(e);
-                }
             }
         }
 
@@ -93,47 +88,38 @@ public class MemoryManager {
             return a;
     }
 
-    public boolean delete_Email(Email mail)
+    public boolean delete_Email (Email mail) throws RuntimeException
     {
         boolean ok = false;
         String path;
         File receiver;
         for (String dest:mail.getReceivers()) {
+            if(!accounts.contains(dest)) throw new RuntimeException("Not found");
             path = dest + "/InArrivo/" + mail.getID();
             receiver = new File(dirpath + "/" + path);
 
             if (receiver.exists()) {
                 receiver.delete();
                 ok=true;
-                model.setLog(model.getLog() + "Email  " + mail.getID() + " eliminata da casella InArrivo di " + dest + "\n");
             } else {
                 path = dest + "/Ricevute/" + mail.getID();
                 receiver = new File(dirpath + "/" + path);
                 if (receiver.exists()) {
                     receiver.delete();
                     ok=true;
-                    model.setLog(model.getLog() + "Email  " + mail.getID() + " eliminata da casella Ricevute di " + dest + "\n");
-                }else model.setLog(model.getLog() + "Email  " + mail.getID() + " non trovata in " + dest + "\n");
+                }
             }
         }
         return ok;
     }
 
-    public boolean move_email(String ID,String dest){
+    public void move_email(String ID,String dest){
         String path1 = dirpath+"/"+ dest + "/InArrivo/" + ID;
         String path2 = dirpath+"/"+ dest + "/Ricevute/" + ID;
 
         File sourceFile = new File(path1);
         File targetFile = new File(path2);
-
-
-        if (sourceFile.renameTo(targetFile)) {
-            model.setLog(model.getLog() +"Email  " + ID + " inviata a " + dest + "\n");
-            return true;
-        } else {
-            System.err.println("Nessuna mail con ID "+ID+" in arrivo a "+dest);
-            return false;
-        }
+        sourceFile.renameTo(targetFile);
     }
 
     public ArrayList<Email> getMails(boolean bool,String dest) throws IOException, ClassNotFoundException {
@@ -166,11 +152,5 @@ public class MemoryManager {
         if (reader!=null)
             reader.close();
     }
-
-
-    public void close(){
-        ID_Seq.close();
-    }
-
 
 }
